+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `{copy_|blit}_[scaled]_rect` to `grixy::ops` (and _unchecked_ variants)
+- Added `GridDraw` to `grixy::ops`
+- Added `grid::ops::blend` for example blend functions
 - Added `grixy::prelude` module for common imports
 
 ### Changed